python/requirements.txt
DateTime==4.3
pkg-resources==0.0.0
PyMySQL==0.9.3
python-dateutil==2.8.1
python-dotenv==0.10.3
python-http-client==3.2.1
pytz==2019.3
sendgrid==6.1.0
six==1.13.0
zope.interface==4.7.1
